Make time constraint into its own class
