+++ src/app/modules/shared/models/constraints/concept-constraint.ts
@@ -1,7 +1,7 @@
 import {Constraint} from './constraint';
 import {Concept} from "../concept";
 import {Value} from "../value";
-import {DateOperatorState} from "../../../data-selection/constraint-components/concept-constraint/date-operator-state";
+import {TimeConstraint} from "./time-constraint";
 
 export class ConceptConstraint implements Constraint {
 
@@ -9,9 +9,7 @@ export class ConceptConstraint implements Constraint {
   private _concept:Concept;
   private _values: Value[];
   applyDateConstraint: boolean = false;
-  dateOperator: DateOperatorState = DateOperatorState.BETWEEN;
-  date1: Date = new Date();
-  date2: Date = new Date();
+  timeConstraint: TimeConstraint = new TimeConstraint();
 
   constructor() {
     this._type = 'ConceptConstraint';
@@ -57,7 +55,7 @@ export class ConceptConstraint implements Constraint {
     }
 
     if (this.applyDateConstraint) {
-      args.push(this.buildDateConstraintQueryObject());
+      args.push(this.timeConstraint.toQueryObject());
     }
 
     return {
@@ -73,45 +71,4 @@ export class ConceptConstraint implements Constraint {
     return 'Concept';
   }
 
-  /** Builds a query object for the date constraint.
-   * @returns {Object}
-   */
-  buildDateConstraintQueryObject():Object {
-    // Operator
-    let operator = {
-      [DateOperatorState.BETWEEN]: "<-->",
-      [DateOperatorState.NOT_BETWEEN]: "<-->", // we'll negate it later
-      [DateOperatorState.BEFORE]: "<-",
-      [DateOperatorState.AFTER]: "->"
-    }[this.dateOperator];
-
-    // Values (dates)
-    let values = [this.date1.toISOString()];
-    if (this.dateOperator == DateOperatorState.BETWEEN ||
-        this.dateOperator == DateOperatorState.NOT_BETWEEN) {
-      values.push(this.date2.toISOString());
-    }
-
-    // Construct the date constraint
-    let dateConstraint:Object = {
-      type: "time",
-      field: {
-        dimension: "start time",
-        fieldName: "startDate",
-        type: "DATE"
-      },
-      operator: operator,
-      values: values
-    };
-
-    // Wrap date constraint in a negation if required
-    if (this.dateOperator == DateOperatorState.NOT_BETWEEN) {
-      dateConstraint = {
-        type: "negation",
-        arg: dateConstraint
-      };
-    }
-
-    return dateConstraint;
-  }
 }
